refactor(core): split edges/extract god-module — extract python.rs
extract.rs was 2619 lines / 97 fns — the second-largest god-module and
the densest CST edge-extraction logic (6 design/bug-pattern repo
memories). Convert it to a directory module and lift the fully
self-contained Python cluster:

- extract/python.rs: python_edges (the Language::Python arm) +
  emit_python_type_refs + the python_* closed-recognizer helpers +
  the python_edge_tests module (46 tests). Verified self-contained —
  no python_* helper is called outside the cluster.
- extract/mod.rs: the entry points, the rust/typescript/kotlin/c_like
  extractors, the shared edge-building helpers, and the rust dispatch
  subsystem. `syntactic_edges` now dispatches via python::python_edges;
  python_edges tightened to pub(super).

Pure code movement — no extraction logic changed; the invariants the
repo memories guard are preserved. extract/mod.rs drops to 1490 lines.
cargo +nightly fmt, clippy --all-targets, and nextest -p rag-rat-core
(868 tests, incl. the 46 moved python_edge_tests + the dispatch tests)
all green.
